+++ src/types.ts
@@ -6,3 +6,10 @@ export type Game = {
   blockTimestamp: string;
   blockNumber: string;
 };
+
+export type Player = {
+  __typename: string;
+  id: string;
+  wins: string;
+  losses: string;
+};
