--- tests/sample.test.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-
-describe("CardGrid", () => {
-  // TODO: render comp
-  expect(true).toBeTruthy();
-});
