--- types/player.ts
@@ -1,6 +1,0 @@
-export type Player = {
-  __typename: string;
-  id: string;
-  wins: string;
-  losses: string;
-};
